@@ -1,63 +1,57 @@
 {
   "name": "client",
+  "description": "Web client using Sveltekit, Routify, SMUI and WindiCSS for theming, with support of GraphQL through URQL.",
   "version": "0.0.0",
-  "description": "Web client using Svelte, Routify, Svelte-Materialify and WindiCSS for theming, with support of GraphQL through URQL.",
+  "type": "module",
   "private": true,
   "scripts": {
-    "dev": "npm-run-all -p gql:watch dev:no-gql",
-    "dev:no-gql": "npm-run-all -p dev:routify dev:vite",
-    "dev:routify": "routify",
-    "dev:vite": "vite --host",
-    "build": "npm-run-all -s build:routify build:vite",
-    "build:routify": "routify -b",
-    "build:vite": "vite build",
-    "serve:vite": "vite preview",
-    "validate": "npm-run-all validate:*",
-    "validate:svelte": "svelte-check --diagnostic-sources \"js,svelte\"",
+    "dev": "svelte-kit dev --port 5000",
+    "build": "svelte-kit build",
+    "preview": "svelte-kit preview --port 5000",
+    "validate": "npm-run-all -p validate:*",
+    "validate:svelte": "svelte-check --diagnostic-sources \"js,ts,svelte\"",
     "validate:ts": "tsc --noEmit",
-    "preview": "vite preview --port 5004",
-    "lint": "npm-run-all lint:*",
+    "lint": "npm-run-all -p lint:*",
     "lint:src": "eslint \"src/**/*.{t,j}s\"",
-    "lint:css": "stylelint \"**/*.{css,svelte,html}\"",
-    "gql": "graphql-codegen -r dotenv/config --config codegen.yml dotenv_config_path=.env.local",
-    "gql:watch": "graphql-codegen -r dotenv/config --config codegen.yml dotenv_config_path=.env.local --watch"
+    "lint:css": "stylelint \"src/**/*.{css,svelte,html}\"",
+    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. .",
+    "smui-theme": "npm-run-all -p smui-theme:*",
+    "smui-theme:light": "smui-theme compile static/theme/smui.css -i src/theme/light",
+    "smui-theme:dark": "smui-theme compile static/theme/smui-dark.css -i src/theme/dark",
+    "prepare": "npm run smui-theme"
   },
   "dependencies": {
     "@urql/svelte": "1.3.3",
-    "axios": "0.25.0",
-    "graphql": "15.8.0",
+    "graphql": "16.3.0",
     "graphql-ws": "5.5.5",
-    "minimum-delayer": "0.2.1"
+    "minimum-delayer": "0.2.1",
+    "workbox-precaching": "6.5.1",
+    "workbox-routing": "6.5.1",
+    "workbox-strategies": "6.5.1",
+    "workbox-window": "6.5.1"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "2.6.1",
-    "@graphql-codegen/typed-document-node": "2.2.3",
-    "@graphql-codegen/typescript": "2.4.3",
-    "@graphql-codegen/typescript-operations": "2.3.0",
+    "@material/theme": "13.0.0",
+    "@material/typography": "13.0.0",
     "@mdi/js": "6.5.95",
-    "@roxi/routify": "2.18.5",
-    "@sveltejs/vite-plugin-svelte": "1.0.0-next.40",
-    "@tsconfig/svelte": "3.0.0",
+    "@roxi/routify": "3.0.0-next.83",
+    "@sveltejs/adapter-netlify": "1.0.0-next.51",
+    "@sveltejs/kit": "1.0.0-next.301",
     "@v-ed/tsconfig": "0.1.2",
     "dotenv": "16.0.0",
-    "postcss": "8.4.5",
-    "postcss-html": "1.3.0",
-    "sass": "1.49.7",
+    "eslint-plugin-svelte3": "3.4.1",
+    "prettier-plugin-svelte": "2.6.0",
+    "sass": "1.49.9",
+    "smui-theme": "6.0.0-beta.15",
     "svelte": "3.46.4",
-    "svelte-check": "2.4.3",
-    "svelte-materialify": "0.3.11",
-    "svelte-preprocess": "4.10.3",
-    "svelte-windicss-preprocess": "4.2.6",
-    "typescript": "4.5.5",
-    "urql": "2.1.3",
-    "vite": "2.8.0",
-    "vite-plugin-html": "2.1.2",
-    "vite-plugin-pwa": "0.11.13",
-    "vite-plugin-windicss": "1.7.0",
-    "vite-tsconfig-paths": "3.4.0",
-    "windicss": "3.4.3"
-  },
-  "routify": {
-    "routifyDir": ".routify"
+    "svelte-check": "2.4.6",
+    "svelte-material-ui": "6.0.0-beta.15",
+    "svelte-preprocess": "4.10.4",
+    "tslib": "2.3.1",
+    "typescript": "4.6.2",
+    "urql": "2.2.0",
+    "vite-plugin-windicss": "1.8.3",
+    "vite-tsconfig-paths": "3.4.1",
+    "windicss": "3.5.1"
   }
 }
