@@ -1,18 +1,45 @@
-/** @type {import('svelte-preprocess').default} */
-// @ts-ignore
-const sveltePreprocess = require('svelte-preprocess');
-// const { windi: windicssPreprocess } = require('svelte-windicss-preprocess');
+import adapter from '@sveltejs/adapter-netlify';
+import preprocess from 'svelte-preprocess';
+import VitePluginRoutify from '@roxi/routify/vite-plugin';
+import VitePluginTsConfigPaths from 'vite-tsconfig-paths';
+import VitePluginWindicss from 'vite-plugin-windicss';
 
-const preprocessors = {
-	svelte: sveltePreprocess({
-		scss: {
-			includePaths: ['theme'],
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	// Consult https://github.com/sveltejs/svelte-preprocess
+	// for more information about preprocessors
+	preprocess: [preprocess()],
+
+	kit: {
+		adapter: adapter(),
+
+		vite: {
+			plugins: [
+				VitePluginWindicss(),
+				VitePluginTsConfigPaths({
+					loose: true,
+				}),
+				VitePluginRoutify({ routesDir: 'src/pages' }),
+			],
+			resolve: {
+				dedupe: ['svelte'],
+			},
+
+			optimizeDeps: {
+				exclude: ['@urql/svelte'],
+			},
+
+			// @ts-ignore
+			ssr: {
+				noExternal: [/^@material\//, /^@smui(?:-extra)?\//],
+			},
 		},
-	}),
-	// windicss: windicssPreprocess(),
+
+		browser: {
+			router: false,
+			hydrate: true,
+		},
+	},
 };
 
-module.exports = {
-	preprocess: Object.values(preprocessors),
-	preprocessors,
-};
+export default config;
