@@ -1,3 +1,3 @@
-<!-- routify:options title="help from test 2" -->
+<!-- routify:meta title="help from test 2" -->
 
 <h1>Hi from test 2!</h1>
